+++ simulations/esercizio_4/statistics.h
@@ -4,6 +4,6 @@
 using namespace std;
 
 void blocked_stats(vector<double> &, vector<double> &, int, string);
-float error(vector<double>, vector<double>, int);
+double error(vector<double>, vector<double>, int);
 
 #endif
